requirements.txt
# Core dependencies
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
schedule>=1.2.0

# Data providers
fredapi>=0.5.0
yfinance>=0.2.18

# Exchange APIs
ccxt>=4.1.0
binance-connector>=3.7.0

# Async support
aiohttp>=3.8.0
asyncio-throttle>=1.0.2

# Utilities
pydantic>=2.0.0
python-dateutil>=2.8.0
pytz>=2023.3

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# Optional for advanced features
scipy>=1.11.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0